feat: add git install and pull to pdoc workflow
This update modifies the pdoc workflow by adding git to the list of installed dependencies and incorporating a step to perform a `git pull`. The inclusion of git ensures that the latest version of the codebase is always used for generating documentation, enhancing the accuracy and relevance of the output. The change acknowledges the critical role of version control in continuous integration environments and aims to reduce discrepancies between the documentation and the codebase.

By ensuring the documentation is always in sync with the latest code, this adjustment helps to prevent confusion and fosters better understanding and usage of the software by end users and contributors alike.