feat(plankapy): add installation and usage guidelines
Revised the README.md to include comprehensive sections on installation and usage, providing a clear path for users to get started with plankapy. Additionally, updates were made to highlight the fork's aim of making plankapy pip installable and ensuring compatibility with a wider range of Python versions by adjusting the requires-python version from >=3.10 to >=3.8 in pyproject.toml. This expansion not only increases accessibility for developers using older Python versions but also aligns with our goal to make the project more inclusive and user-friendly.